requirements.txt
django==2.2
mysqlclient
pillow
django-pure-pagination
redis
django-simple-captcha